requirements.txt
PyMuPDF
python-docx
